add ut for publishEvent
