+++ sdk/java-sdk/sdk/src/test/java/io/mosn/layotto/v1/SubscriberTestWithRealServer.java
@@ -18,6 +18,7 @@ import com.google.protobuf.ByteString;
 import com.google.protobuf.Empty;
 import io.grpc.ManagedChannel;
 import io.grpc.ManagedChannelBuilder;
+import io.mosn.layotto.v1.mock.MySubscriber;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
